wip
fixed routeservice handler tests. The test setup was wrong. We also
altered the test url to contain a newline in the query parameter to
validate that our fix still works

also changed other tests to no longer QueryUnescape the forwardedUrl in
setup because this is what was causing the problem.

Co-authored-by: Kauana dos Santos <kdossantos@pivotal.io>