example for forwarding http.Server errors
This work shows how we can capture the errors from http.Server and use
our own logger to write the logs. The only downside so far is that it
will greatly change the format of the logs and the log location. This
does make it more difficult to read and could break tooling that looks
at stderr.log

Co-authored-by: Amelia Downs <adowns@pivotal.io>